backend/requirements.txt
# RAG backend: bilingual Startup Proclamation
langchain
langchain-core
langchain-community
langchain-chroma
langchain-openai
langchain-google-genai
fastapi
pydantic
python-dotenv
openai
pymupdf
PyYAML
uvicorn[standard]
